Remove controlWrapper from inputs
