+++ addon/templates/components/em-input.hbs
@@ -1,9 +1,8 @@
 {{#em-form-group model=model property=property label=label form=form
   elementClass=elementClass controlWrapper=controlWrapper labelClass=labelClass
-  label=label yieldInLabel=yieldInLabel
-  labelWrapperClass=labelWrapperClass labelClass=labelClass
-  help=help controlWrapper=controlWrapper
-  cid=id validationIcons=validationIcons canShowErrors=canShowErrors inputComponent=this}}
+  label=label yieldInLabel=yieldInLabel labelWrapperClass=labelWrapperClass
+  help=help cid=id validationIcons=validationIcons canShowErrors=canShowErrors
+  inputComponent=this}}
   {{input
     type=type
