+++ addon/templates/components/em-text.hbs
@@ -1,8 +1,6 @@
 {{#em-form-group model=model property=property label=label form=form
-  elementClass=elementClass controlWrapper=controlWrapper labelClass=labelClass
-  label=label yieldInLabel=yieldInLabel
-  labelWrapperClass=labelWrapperClass labelClass=labelClass
-  help=help controlWrapper=controlWrapper
+  elementClass=elementClass labelClass=labelClass label=label
+  yieldInLabel=yieldInLabel labelWrapperClass=labelWrapperClass help=help
   cid=id validationIcons=validationIcons canShowErrors=canShowErrors inputComponent=this}}
   {{textarea
     type=type
